# packages/api-server/src/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: HappyCMDB API
  version: 2.0.0
  description: |
    Comprehensive REST API for HappyCMDB - Enterprise Configuration Management Database (CMDB) platform.

    HappyCMDB v2.0 features:
    - Graph-based CI relationships with Neo4j
    - 43 connectors (17 TypeScript + 26 JSON-only)
    - Unified credential system with affinity matching
    - Discovery agent architecture with smart routing
    - AI-powered anomaly detection and impact analysis
    - Identity resolution and reconciliation

    ## Authentication

    All API endpoints (except `/api/v1/auth/login`) require authentication via:
    - **JWT Token**: Include `Authorization: Bearer <token>` header
    - **API Key**: Include `X-API-Key: <key>` header

    ## Rate Limiting

    Rate limits apply per authentication token/API key:
    - **Standard**: 1000 requests/hour
    - **Burst**: 100 requests/minute

    Rate limit headers included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: HappyCMDB Support
    url: http://localhost:8080
    email: support@happycmdb.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://staging-api.happycmdb.io/api/v1
    description: Staging server
  - url: https://api.happycmdb.io/api/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /cis:
    get:
      tags: [Configuration Items]
      summary: List all CIs
      description: Retrieve all Configuration Items with optional filtering
      operationId: getAllCIs
      parameters:
        - name: _type
          in: query
          schema:
            $ref: '#/components/schemas/CIType'
        - name: _status
          in: query
          schema:
            $ref: '#/components/schemas/CIStatus'
        - name: _environment
          in: query
          schema:
            $ref: '#/components/schemas/Environment'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of CIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CI'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CIForbidden'

    post:
      tags: [Configuration Items]
      summary: Create new CI
      description: Create a new Configuration Item
      operationId: createCI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CIInput'
            examples:
              server:
                summary: Create a server CI
                value:
                  _id: srv-prod-web-01
                  name: Production Web Server 01
                  _type: server
                  _status: active
                  environment: production
                  metadata:
                    hostname: web01.example.com
                    ip_address: 192.168.1.10
                    os: Ubuntu 22.04
                    cpu_cores: 8
                    memory_gb: 32
              application:
                summary: Create an application CI
                value:
                  _id: app-ecommerce-api
                  name: E-Commerce API
                  _type: application
                  _status: active
                  environment: production
                  metadata:
                    version: 2.3.1
                    port: 8080
                    framework: Node.js
      responses:
        '201':
          description: CI created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/CIBodyBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CIForbidden'
        '409':
          description: CI with this ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /relationships:
    get:
      tags: [Relationships]
      summary: List relationships
      description: Retrieve relationships with optional filtering
      operationId: listRelationships
      parameters:
        - name: _type
          in: query
          schema:
            $ref: '#/components/schemas/RelationshipType'
        - name: _from_id
          in: query
          schema:
            type: string
          description: Filter by source CI ID
        - name: _to_id
          in: query
          schema:
            type: string
          description: Filter by target CI ID
        - name: _ci_id
          in: query
          schema:
            type: string
          description: Filter by any CI ID (source or target)
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of relationships
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Relationship'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Relationships]
      summary: Create relationship
      description: Create a relationship between two CIs
      operationId: createRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipInput'
            examples:
              dependency:
                summary: Application depends on database
                value:
                  _from_id: app-ecommerce-api
                  _to_id: db-postgres-prod
                  _type: DEPENDS_ON
                  properties:
                    connection_string: postgresql://db-postgres-prod:5432/ecommerce
              hosting:
                summary: Server hosts application
                value:
                  _from_id: srv-prod-web-01
                  _to_id: app-ecommerce-api
                  _type: HOSTS
      responses:
        '201':
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relationship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: One or both CIs not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /auth/api-keys endpoint
  parameters:
...
    Limit:
      name: _limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: _offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation Error
            message: Invalid CI type provided
            details:
              - field: _type
                error: must be one of [server, virtual-machine, ...]
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired JWT token
...
    CIForbidden:
      description: >-
        The authenticated user has no organization: every /cis endpoint is
        tenant-scoped and rejects the request before any data access. Mutations
        also return 403 without the `write` permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MiddlewareError'
          example:
            _error: Forbidden
            _message: Organization claim required
    CIBodyBadRequest:
      description: >-
        Invalid CI body. `organization_id` is never accepted in a body (the
        organization always comes from the token) and is rejected with 400.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MiddlewareError'
          example:
            _success: false
            _error: Validation Error
            _message: '"organization_id" is not allowed'
            _details:
              - _field: organization_id
                _message: '"organization_id" is not allowed'
                _type: any.unknown
...
  schemas:
...
    CI:
      type: object
      required: [_id, name, _type, _status, _created_at, _updated_at, _discovered_at]
      properties:
        _id:
          type: string
          description: Unique identifier
          example: srv-prod-web-01
        external_id:
          type: string
          description: External system ID (e.g., AWS instance ID)
          example: i-0123456789abcdef0
        name:
          type: string
          description: Human-readable name
          example: Production Web Server 01
        _type:
          $ref: '#/components/schemas/CIType'
        _status:
          $ref: '#/components/schemas/CIStatus'
        environment:
          $ref: '#/components/schemas/Environment'
        _created_at:
          type: string
          format: date-time
          description: When CI was created in CMDB
        _updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        _discovered_at:
          type: string
          format: date-time
          description: When CI was discovered
        _metadata:
          type: object
          additionalProperties: true
          description: Additional attributes and metadata
          example:
            hostname: web01.example.com
            ip_address: 192.168.1.10
            os: Ubuntu 22.04
            cpu_cores: 8
            memory_gb: 32
    CIInput:
      type: object
      required: [_id, name, _type]
      properties:
        _id:
          type: string
        external_id:
          type: string
        name:
          type: string
        _type:
          $ref: '#/components/schemas/CIType'
        status:
          $ref: '#/components/schemas/CIStatus'
        environment:
          $ref: '#/components/schemas/Environment'
        discovered_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    CIType:
      type: string
      enum:
        - server
        - virtual-machine
        - container
        - application
        - service
        - database
        - network-device
        - storage
        - load-balancer
        - cloud-resource
        - alert
        - vulnerability
        - software
        - detection
        - incident
        - hardware-component
        - mobile-device
        - policy
        - configuration
        - user
        - group
        - organizational-unit
        - collection
        - update
    CIStatus:
      type: string
      enum: [active, inactive, maintenance, decommissioned]
    Environment:
      type: string
      enum: [production, staging, development, test]
    Relationship:
      type: object
      required: [_from_id, _to_id, _type]
      properties:
        _from_id:
          type: string
          description: Source CI ID
        _to_id:
          type: string
          description: Target CI ID
        _type:
          $ref: '#/components/schemas/RelationshipType'
        properties:
          type: object
          additionalProperties: true
          description: Additional relationship metadata
    RelationshipInput:
      type: object
      required: [_from_id, _to_id, _type]
      properties:
        _from_id:
          type: string
        _to_id:
          type: string
        _type:
          $ref: '#/components/schemas/RelationshipType'
        properties:
          type: object
          additionalProperties: true
    RelationshipType:
      type: string
      enum:
        - DEPENDS_ON
        - HOSTS
        - CONNECTS_TO
        - USES
        - OWNED_BY
        - PART_OF
        - LOCATED_IN
        - DEPLOYED_ON
        - BACKED_UP_BY
        - DETECTED_ON
        - AFFECTS
        - INSTALLED_ON
        - CONTAINS
        - ASSIGNED_TO
        - APPLIES_TO
        - MEMBER_OF
        - REQUIRED_BY
        - HOSTED_ON

    # =========================================================================
    # AUTHENTICATION
    # =========================================================================
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type/code
          example: Validation Error
        message:
          type: string
          description: Human-readable error message
          example: Invalid CI type provided
        details:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
              error:
                type: string
    MiddlewareError:
      type: object
      description: >-
        Error body produced by the auth and request-validation middleware
        (underscore-prefixed keys), e.g. the /cis 403 and body-validation 400.
      required: [_error, _message]
      properties:
        _success:
          type: boolean
          example: false
        _error:
          type: string
          example: Validation Error
        _message:
          type: string
          example: '"organization_id" is not allowed'
        _details:
          type: array
          items:
            type: object
            properties:
              _field:
                type: string
              _message:
                type: string
              _type:
                type: string